+++ src/codegen/templates/task/channels/_ao_channel.py.mako
@@ -1,24 +1,25 @@
 <%
     from codegen.utilities.attribute_helpers import get_attributes, get_enums_used
     from codegen.utilities.text_wrappers import wrap
-    attributes = get_attributes(data, "DIChannel")
+    attributes = get_attributes(data, "AOChannel")
     enums_used = get_enums_used(attributes)
 %>\
 # Do not edit this file; it was automatically generated.
 
-from nidaqmx._task_modules.channels.channel import Channel
+from nidaqmx.scale import _ScaleAlternateConstructor
+from nidaqmx.task.channels._channel import Channel
 from nidaqmx.constants import (
     ${', '.join([c for c in enums_used]) | wrap(4, 4)})
 
 
-class DIChannel(Channel):
+class AOChannel(Channel):
     """
-    Represents one or more digital input virtual channels and their properties.
+    Represents one or more analog output virtual channels and their properties.
     """
     __slots__ = ()
 
     def __repr__(self):
-        return f'DIChannel(name={self._name})'
+        return f'AOChannel(name={self._name})'
 
 <%namespace name="property_template" file="/property_template.py.mako"/>\
 %for attribute in attributes:
