+++ src/codegen/templates/task/channels/_channel.py.mako
@@ -6,7 +6,7 @@
 %>\
 # Do not edit this file; it was automatically generated.
 
-import nidaqmx._task_modules.channels  # circular import: Channel._factory uses derived classes
+import nidaqmx.task.channels  # circular import: Channel._factory uses derived classes
 from nidaqmx.system.physical_channel import _PhysicalChannelAlternateConstructor
 from nidaqmx.utils import flatten_channel_string, unflatten_channel_string
 from nidaqmx.constants import (
@@ -99,7 +99,7 @@ class Channel:
                 or physical name of a channel.
             interpreter (BaseInterpreter): Specifies the interpreter instance.
         Returns:
-            nidaqmx._task_modules.channels.channel.Channel:
+            nidaqmx.task.channels.Channel:
 
             Indicates an object that represents the specified channel.
         """
@@ -108,22 +108,22 @@ class Channel:
         channel_type = ChannelType(chan_type)
 
         if channel_type == ChannelType.ANALOG_INPUT:
-            return nidaqmx._task_modules.channels.AIChannel(
+            return nidaqmx.task.channels.AIChannel(
                 task_handle, virtual_or_physical_name, interpreter)
         elif channel_type == ChannelType.ANALOG_OUTPUT:
-            return nidaqmx._task_modules.channels.AOChannel(
+            return nidaqmx.task.channels.AOChannel(
                 task_handle, virtual_or_physical_name, interpreter)
         elif channel_type == ChannelType.COUNTER_INPUT:
-            return nidaqmx._task_modules.channels.CIChannel(
+            return nidaqmx.task.channels.CIChannel(
                 task_handle, virtual_or_physical_name, interpreter)
         elif channel_type == ChannelType.COUNTER_OUTPUT:
-            return nidaqmx._task_modules.channels.COChannel(
+            return nidaqmx.task.channels.COChannel(
                 task_handle, virtual_or_physical_name, interpreter)
         elif channel_type == ChannelType.DIGITAL_INPUT:
-            return nidaqmx._task_modules.channels.DIChannel(
+            return nidaqmx.task.channels.DIChannel(
                 task_handle, virtual_or_physical_name, interpreter)
         elif channel_type == ChannelType.DIGITAL_OUTPUT:
-            return nidaqmx._task_modules.channels.DOChannel(
+            return nidaqmx.task.channels.DOChannel(
                 task_handle, virtual_or_physical_name, interpreter)
 
     @property
