+++ src/codegen/templates/task/collections/_ai_channel_collection.py.mako
@@ -8,7 +8,7 @@
 
 import numpy
 
-from nidaqmx._task_modules.channels.ai_channel import AIChannel
+from nidaqmx.task.channels._ai_channel import AIChannel
 from nidaqmx.task.collections._channel_collection import ChannelCollection
 from nidaqmx.utils import unflatten_channel_string
 %if enums_used:
@@ -37,7 +37,7 @@ class AIChannelCollection(ChannelCollection):
             name_to_assign_to_channel (Optional[str]): Specifies a name to
                 assign to the virtual channel this method creates.
         Returns:
-            nidaqmx._task_modules.channels.ai_channel.AIChannel:
+            nidaqmx.task.channels.AIChannel:
 
             Specifies the newly created AIChannel object.
         """
