+++ src/codegen/templates/task/collections/_ao_channel_collection.py.mako
@@ -6,7 +6,7 @@
 %>\
 # Do not edit this file; it was automatically generated.
 
-from nidaqmx._task_modules.channels.ao_channel import AOChannel
+from nidaqmx.task.channels._ao_channel import AOChannel
 from nidaqmx.task.collections._channel_collection import ChannelCollection
 from nidaqmx.utils import unflatten_channel_string
 %if enums_used:
@@ -35,7 +35,7 @@ class AOChannelCollection(ChannelCollection):
             name_to_assign_to_channel (Optional[str]): Specifies a name to
                 assign to the virtual channel this method creates.
         Returns:
-            nidaqmx._task_modules.channels.ao_channel.AOChannel: 
+            nidaqmx.task.channels.AOChannel: 
             
             Specifies the newly created AOChannel object.
         """
