+++ src/codegen/templates/task/collections/_ci_channel_collection.py.mako
@@ -6,7 +6,7 @@
 %>\
 # Do not edit this file; it was automatically generated.
 
-from nidaqmx._task_modules.channels.ci_channel import CIChannel
+from nidaqmx.task.channels._ci_channel import CIChannel
 from nidaqmx.task.collections._channel_collection import ChannelCollection
 from nidaqmx.utils import unflatten_channel_string
 %if enums_used:
@@ -35,7 +35,7 @@ class CIChannelCollection(ChannelCollection):
             name_to_assign_to_channel (Optional[str]): Specifies a name to 
                 assign to the virtual channel this method creates.
         Returns:
-            nidaqmx._task_modules.channels.ci_channel.CIChannel: 
+            nidaqmx.task.channels.CIChannel: 
             
             Specifies the newly created CIChannel object.
         """
