+++ src/codegen/templates/task/collections/_di_channel_collection.py.mako
@@ -6,7 +6,7 @@
 %>\
 # Do not edit this file; it was automatically generated.
 
-from nidaqmx._task_modules.channels.di_channel import DIChannel
+from nidaqmx.task.channels._di_channel import DIChannel
 from nidaqmx.task.collections._channel_collection import ChannelCollection
 from nidaqmx.utils import unflatten_channel_string
 %if enums_used:
@@ -38,7 +38,7 @@ class DIChannelCollection(ChannelCollection):
             name_to_assign_to_lines (Optional[str]): Specifies a name to 
                 assign to the virtual channel this method creates.
         Returns:
-            nidaqmx._task_modules.channels.di_channel.DIChannel: 
+            nidaqmx.task.channels.DIChannel: 
             
             Specifies the newly created DIChannel object.
         """
