+++ src/codegen/templates/task/collections/_do_channel_collection.py.mako
@@ -6,7 +6,7 @@
 %>\
 # Do not edit this file; it was automatically generated.
 
-from nidaqmx._task_modules.channels.do_channel import DOChannel
+from nidaqmx.task.channels._do_channel import DOChannel
 from nidaqmx.task.collections._channel_collection import ChannelCollection
 from nidaqmx.utils import unflatten_channel_string
 %if enums_used:
@@ -38,7 +38,7 @@ class DOChannelCollection(ChannelCollection):
             name_to_assign_to_lines (Optional[str]): Specifies a name to 
                 assign to the virtual channel this method creates.
         Returns:
-            nidaqmx._task_modules.channels.do_channel.DOChannel: 
+            nidaqmx.task.channels.DOChannel: 
             
             Specifies the newly created DOChannel object.
         """
