--- src/handwritten/_task_modules/channels/__init__.py
@@ -1,11 +1,0 @@
-__author__ = 'National Instruments'
-__all__ = ['channel']
-
-
-from nidaqmx._task_modules.channels.channel import Channel
-from nidaqmx._task_modules.channels.ai_channel import AIChannel
-from nidaqmx._task_modules.channels.ao_channel import AOChannel
-from nidaqmx._task_modules.channels.ci_channel import CIChannel
-from nidaqmx._task_modules.channels.co_channel import COChannel
-from nidaqmx._task_modules.channels.di_channel import DIChannel
-from nidaqmx._task_modules.channels.do_channel import DOChannel
