+++ src/handwritten/task/collections/_channel_collection.py
@@ -1,6 +1,6 @@
 from collections.abc import Sequence
 
-from nidaqmx._task_modules.channels.channel import Channel
+from nidaqmx.task.channels._channel import Channel
 from nidaqmx.errors import DaqError
 from nidaqmx.error_codes import DAQmxErrors
 from nidaqmx.utils import unflatten_channel_string, flatten_channel_string
@@ -50,7 +50,7 @@ class ChannelCollection(Sequence):
                 - slice: Range of the indexes/positions of virtual channels in
                     the collection.
         Returns:
-            nidaqmx._task_modules.channels.channel.Channel: 
+            nidaqmx.task.channels.Channel: 
             
             Indicates a channel object representing the subset of virtual
             channels indexed.
@@ -96,7 +96,7 @@ class ChannelCollection(Sequence):
     @property
     def all(self):
         """
-        :class:`nidaqmx._task_modules.channels.channel.Channel`:
+        :class:`nidaqmx.task.channels.Channel`:
             Specifies a channel object that represents the entire list of 
             virtual channels on this channel collection.
         """
